requirements.txt
Flask~=2.0.1
Flask-SQLAlchemy~=2.5.1
beautifulsoup4~=4.10.0
psycopg2~=2.9.1
python-dotenv~=0.19.0
requests~=2.26.0
